requirements.txt
# ---- Core ----
fastapi
uvicorn[standard]
python-dotenv

# ---- LangChain & RAG ----
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
langchain-huggingface
langchain-text-splitters

# ---- Embeddings ----
sentence-transformers

# ---- UI (ローカル/公開用) ----
gradio
streamlit

# ---- Utils ----
pydantic>=2,<3

# ---- 互換性のためバージョン固定 ----
numpy<2.0.0
pandas>=2.0.0,<2.1.0
scikit-learn>=1.3.0,<1.4.0

# ---- For PDF/text handling (必要に応じて) ----
tqdm
